Update build instructions [skip ci]
Mention that we require Go 1.11 or higher to support the new Go modules
and mention the env var we need on Travis CI to override the Go 1.5
`vendor` experiment in case that's useful to folks.

Also updated the `go get` and `go build` command lines for clone-less
and cloned approaches to building the tool.