Add a test for ensuring our module config is clean.
This test uses `go mod tidy` to ensure that we are running with the
simplest and cleanest `go mod` config that we can, given our current set
of dependencies. `go mod tidy` should be consistent for the same inputs,
so if we're not changing our (pinned) dependencies, this test should not
fail.

However, it should fail (intentionally) if we change any dependencies or
any versions of dependencies and add new dependencies without removing
the old ones.