Factor out ProcessCommit() and add a test
The test ensures that a commit where the author and committer are different in a
particular aspect (such as the email address used) can still be considered valid
(thus ensuring that they're checked independetly.

This also incrementally reduces the complexity of the extremely large function
ProcessOrgRepo() which was due for refactoring anyway.