Add support for externally-managed CLAs
This addresses the feature request in issue #27 by enabling incremental
migration to this tool (if using something else), or away from this tool
(if switching to something else), without having to wait until everyone
is added to the new system and removed from the old system.

This change adds support for a new CLA signers config field `external`
where you can list contributors in the same way as in the main CLA
signers block (people, bots, and companies which include groups of
people), as well as a special config setting `unknown_as_external` which
can be set to `true` to enable treating any GitHub username not
explicitly listed in the CLA signers file to be treated as
externally-managed contributors.

In the presence of a `[cla: external]` label on the repo, if a commit is
found to be matching the external records (or the wildcard), the entire
PR will be marked with that label, under the expectation that an
external tool will be used for the CLA verification, and CRBot will
simply step aside.