Use `stretchr/testify` for cleaner asserts
Traditionally, Go tests require writing `if not X then fail()` when what
we really want to write is `assert X`. This leads to hard-to-read code
since the condition is inverted, and very verbose if statements with
boilerplate code to output a message via `t.Logf()` followed by
`t.Fail()` on a separate line.

This library simplifies asserts to just a single line, makes it natural
to write and easy to read, without losing the (optional) addition of a
human-readable error message.

There's certainly a cost to having an additional dependency (and its
transitive dependencies), but as I was already considering writing a
very similar library to improve the testing in this project, this is
welcome news to not have to do that myself.