Move secrets to a separate flag/file from config
This allows us to make the config settings public, without also
disclosing the secrets (auth token) file. And the config file itself is
actually optional, as you can specify or override all the settings via
flags.

Also took the opportunity to refactor and simplify some code to remove
duplication, and added custom help for `-help` as well as error output
to clarify that both YAML and JSON file types are supported.

Fixes #47