Fix IsExternal() to handle partially-external PRs
This change fixes the case where one of (author, committer) has signed
the CLA, while the other does not, but the "treat unknown as external"
flag is on, which means `IsExternal()` should return `true` and this
should be handled by the external CLA verification tool.

Previously, `IsExternal()` returned `false` in this case, causing an
error by marking the PR as being non-compliant when it should have been
marked as being handled externally.