Build with Go 1.14, drop Go source build
The prebuilt versions are so much faster (~40-50 sec.) than the source build
(~3.5 min) that it is not worth it, especially since we're not using or relying
on any features that are being added/removed on the Go trunk. This saves us a
bit of time waiting for tests to turn green on PRs prior to merge.

Additional minor fixes:

* s/Google Inc./Google LLC/
* removed deprecated `sudo: false` setting